+++ HISTORY.rst
@@ -5,6 +5,7 @@ History
 Unreleased / master
 -------------------
 * ``ispyb.job``: bug fix - sweeps may start counting at 0
+* Remove deprecated ``ispyb.model``
 
 6.13.0 (2022-11-03)
 -------------------